requirements.txt
aioredis==2.0.1
async-timeout==4.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
cleo==1.0.0a4
click==8.0.4
crashtest==0.3.1
Deprecated==1.2.13
Flask==2.0.3
flask_swagger_ui==4.11.1
hiredis==2.3.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
pptree==3.1
pydantic==1.10.7
pylev==1.4.0
pyparsing==3.0.7
python-dotenv==0.19.2
python-dateutil==2.8.2
python-ulid==1.0.3
redis==4.5.4
redis-om==0.3.1
requests==2.27.1
six==1.16.0
types-redis==4.1.17
types-six==1.16.11
typing-extensions==4.5.0
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
nanoid==2.0.0
PyYAML==6.0.1
